client复用pbft代码；pbft工具函数移入util
