+++ main.go
@@ -4,32 +4,37 @@ import (
 	"github.com/cloudflare/cfssl/log"
 	"github.com/ssbcV2/account"
 	"github.com/ssbcV2/chain"
+	"github.com/ssbcV2/client"
 	"github.com/ssbcV2/common"
 	"github.com/ssbcV2/levelDB"
 	"github.com/ssbcV2/merkle"
 	"github.com/ssbcV2/meta"
 	"github.com/ssbcV2/network"
 	"github.com/ssbcV2/pbft"
+	"github.com/ssbcV2/util"
 	"os"
 )
 
 func main() {
-	//为四个节点生成公私钥
-	pbft.GenRsaKeys()
+	// 为四个节点生成公私钥
+	util.GenRsaKeys()
 
 	if len(os.Args) != 2 {
 		log.Error("输入的参数有误！")
 	}
 	nodeID := os.Args[1]
 	merkle.StatePath = "./levelDB/db/path/statedb/" + nodeID // 账户数据暂时使用单独的数据库存储
-	//数据库连接
+	// 数据库连接
 	levelDB.InitDB(nodeID)
 
 	// 从levelDB读取账户信息（必须在数据库建立连接后，所以不能在init()完成）
 	account.GetFromDisk()
 
 	if nodeID == "client" {
-		pbft.ClientSendMessageAndListen() //启动客户端程序
+		go client.ListenRequest() 	// 启动客户端程序
+		p := pbft.NewPBFT(nodeID, common.ClientToNodeAddr)
+		go p.TcpListen()
+
 		//初始化
 		initBlockChain(nodeID)
 	} else if addr, ok := common.NodeTable[nodeID]; ok {
