+++ util/util.go
@@ -4,6 +4,7 @@ import (
 	"crypto/rand"
 	"github.com/cloudflare/cfssl/log"
 	"math/big"
+	"os"
 )
 
 // 返回一个十位数的随机数，作为msgid
@@ -18,4 +19,20 @@ func GetRandom() int {
 			return int(result.Int64())
 		}
 	}
+}
+
+// 判断文件或文件夹是否存在
+func FileExists(path string) bool {
+	_, err := os.Stat(path)
+	if err != nil {
+		if os.IsExist(err) {
+			return true
+		}
+		if os.IsNotExist(err) {
+			return false
+		}
+		log.Info(err)
+		return false
+	}
+	return true
 }
